+++ rpc/namespaces/ethereum/personal/api.go
@@ -13,7 +13,7 @@ import (
 	"github.com/evmos/evmos/v14/crypto/hd"
 	"github.com/evmos/evmos/v14/types"
 
-	"github.com/tendermint/tendermint/libs/log"
+	"github.com/cometbft/cometbft/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	sdk "github.com/cosmos/cosmos-sdk/types"